requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1.1
django-bootstrap4==24.4
django-filter==24.3
django-phonenumber-field==8.0.0
pillow==10.4.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
